refactor(step-7): move tags.ts to src/operations/tags.ts
tags.ts joins notes.ts and bear-encoding.ts in the operations/ layer.
Its imports now resolve to the correct parent paths (../types,
../logging, ../infra/database) and its sibling bear-encoding.js.
main.ts imports from ./operations/tags.js.
