+++ src/operations/tags.ts
@@ -1,7 +1,8 @@
-import type { BearNote, BearTag } from './types.js';
-import { convertCoreDataTimestamp, decodeTagName } from './operations/bear-encoding.js';
-import { logAndThrow, logger } from './logging.js';
-import { closeBearDatabase, openBearDatabase } from './infra/database.js';
+import type { BearNote, BearTag } from '../types.js';
+import { logAndThrow, logger } from '../logging.js';
+import { closeBearDatabase, openBearDatabase } from '../infra/database.js';
+
+import { convertCoreDataTimestamp, decodeTagName } from './bear-encoding.js';
 
 /**
  * Extracts the display name (leaf) from a full tag path.
